Core: include optimization util::is_stream_insertible
