+++ modules/base/src/processors/distancetransformram.cpp
@@ -50,6 +50,7 @@
 #include <inviwo/core/properties/valuewrapper.h>                       // for PropertySerializat...
 #include <inviwo/core/util/glmvec.h>                                   // for dvec2, size3_t
 #include <inviwo/core/util/staticstring.h>                             // for operator+
+#include <inviwo/core/util/isstreaminsertable.h>                       // for util::is_stream_in...
 #include <modules/base/algorithm/dataminmax.h>                         // for dataMinMax
 #include <modules/base/algorithm/volume/volumeramdistancetransform.h>  // for volumeDistanceTran...
