+++ src/core/CMakeLists.txt
@@ -362,6 +362,7 @@ set(HEADER_FILES
     ${IVW_INCLUDE_DIR}/inviwo/core/util/intersection/raysphereintersection.h
     ${IVW_INCLUDE_DIR}/inviwo/core/util/introspection.h
     ${IVW_INCLUDE_DIR}/inviwo/core/util/inviwosetupinfo.h
+    ${IVW_INCLUDE_DIR}/inviwo/core/util/isstreaminsertable.h
     ${IVW_INCLUDE_DIR}/inviwo/core/util/iterrange.h
     ${IVW_INCLUDE_DIR}/inviwo/core/util/licenseinfo.h
     ${IVW_INCLUDE_DIR}/inviwo/core/util/localetools.h
@@ -727,6 +728,7 @@ set(SOURCE_FILES
     util/indirectiterator.cpp
     util/introspection.cpp
     util/inviwosetupinfo.cpp
+    util/isstreaminsertable.cpp
     util/iterrange.cpp
     util/licenseinfo.cpp
     util/localetools.cpp
